feat: complete refactor of the lib make it so lib is build as a static library that can then be used into different game move all headers file into include/ move game used as test for the lib into example/ enhance Makefile to build lib and game separately
