+++ example/main.cpp
@@ -1,4 +1,4 @@
-#include "engine/core/engine.h"
+#include "core/engine/engine.h"
 #include "game/player.h"
 #include "game/main_scene.h"
 #include "game/ennemy.h"
