+++ include/core/sprite/sprite.h
