+++ include/entities/entity_manager.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <memory>
-#include "engine/entities/entities.h"
+#include "entities/entities.h"
 
 namespace engine {
