+++ src/entities/entities.cpp
@@ -1,4 +1,4 @@
-#include "entities.h"
+#include "entities/entities.h"
 
 namespace engine {
